购物车更新

+++ src/main/java/com/hph/dao/IOperations.java
@@ -22,5 +22,5 @@ public interface IOperations<T extends Serializable> {
 
 	void delete(final T entity);
 
-	void deleteById(final long entityId);
+	void deleteById(final int entityId);
 }
